fix post-refactoring compilation
