+++ opengrok-indexer/src/main/java/org/opengrok/indexer/web/PageConfig.java
@@ -609,7 +609,7 @@ public final class PageConfig {
      */
     public EftarFileReader getEftarReader() {
         if (eftarReader == null || eftarReader.isClosed()) {
-            File f = getEnv().getConfiguration().getDtagsEftar();
+            File f = getEnv().getDtagsEftar();
             if (f == null) {
                 eftarReader = null;
             } else {
