+++ opengrok-web/src/main/java/org/opengrok/web/AuthorizationFilter.java
@@ -75,7 +75,7 @@ public class AuthorizationFilter implements Filter {
                 LOGGER.log(Level.INFO, "Access denied for URI: {0}", httpReq.getRequestURI());
             }
 
-            /**
+            /*
              * Add the request to the statistics. This is called just once for a
              * single request otherwise the next filter will count the same
              * request twice ({@link StatisticsFilter#collectStats}).
@@ -88,7 +88,7 @@ public class AuthorizationFilter implements Filter {
             config.getEnv().getStatistics().addRequestTime("requests_forbidden",
                     System.currentTimeMillis() - processTime);
 
-            if (!config.getEnv().getConfiguration().getForbiddenIncludeFileContent(false).isEmpty()) {
+            if (!config.getEnv().getIncludeFiles().getForbiddenIncludeFileContent(false).isEmpty()) {
                 sr.getRequestDispatcher("/eforbidden").forward(sr, sr1);
                 return;
             }
